requirements.txt
# Core ADK dependencies
google-genai>=0.2.0
google-generativeai>=0.8.0

# Database
sqlalchemy>=2.0.0
sqlite3worker>=1.1.7
psycopg2-binary>=2.9.9

# Data processing and visualization
pandas>=2.1.0
plotly>=5.18.0
matplotlib>=3.8.0

# Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
structlog>=23.2.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
